+++ scripts/filterVcf.pl
@@ -76,11 +76,7 @@ MAIN:
 		die "NORMALIZATION ERROR:$_\n" if($F[4]=~/,/);
 		#$F[6]="." unless($F[6] eq "PASS");
 
-		if($opt{percent}>0)
-		{
-			$F[3]=uc($F[3]);
-        	        $F[4]=uc($F[4]);
-		}
+		($F[3],$F[4])=(uc($F[3]),uc($F[4])) if($opt{percent}>0);
 
 		if($F[8] eq "SM")
 		{